+++ src/loader/auto-placeholder/index.ts
@@ -1,0 +1,1 @@
+export { autoPlaceLoader as default } from './auto-placeholder';