+++ src/loader/css-url-blank-space/index.ts
@@ -1,0 +1,1 @@
+export { cssUrlBlankSpaceLoader as default } from './css-url-blank-space';