+++ src/loader/swipe-component/index.ts
@@ -1,0 +1,1 @@
+export { swipeComponent as default } from './swipe-component';