+++ src/plugin/add-placeholder/index.ts
@@ -1,0 +1,1 @@
+export { AddPlaceHolderPlugin } from './add-placeholder';