+++ src/plugin/copy-dir/index.ts
@@ -1,0 +1,1 @@
+export { CopyDirPlugin } from './copy-dir';