+++ src/plugin/global-this-polyfill/index.ts
@@ -1,0 +1,1 @@
+export { GlobalThisPolyfillPlugin } from './global-this-polyfill';