+++ src/plugin/index.ts
@@ -1,13 +1,17 @@
+export * from './add-placeholder';
 export * from './analyze-deps';
+export * from './copy-dir';
 export * from './dispatch-script';
 export * from './dispatch-vue';
 export * from './fix-npm-package';
 export * from './gen-version';
+export * from './global-this-polyfill';
 export * from './modify-html';
 export * from './rem-to-rpx';
 export * from './replace-app-id';
 export * from './replace-content';
 export * from './save-loader-log';
 export * from './transfer-local-file';
+
 
 export * from '../vite-plugin';