+++ src/webpack/uni-vue-config/config.ts
@@ -1,0 +1,32 @@
+export const DEFAULT_TRANSPILE_DEPENDENCIES = [
+  'pmd-merchant-ui',
+  'press-ui',
+  'press-plus',
+
+  'pmd-aegis',
+  'pmd-app-info',
+  'pmd-config',
+  'pmd-location',
+  'pmd-login',
+  'pmd-network',
+  'pmd-report',
+  'pmd-tools',
+  'pmd-types',
+  'pmd-widget',
+  'pmd-vue',
+  'pmd-jsapi',
+];
+
+export const DEFAULT_ADAPTER_DIRS = [
+  'comm',
+  'common',
+  'component',
+  'live-component',
+  'logic',
+  'local-logic',
+  'local-component',
+  'login',
+  'pages',
+  'static',
+  'node-modules',
+];