+++ src/webpack/uni-vue-config/index.ts
@@ -1,0 +1,1 @@
+export { getUniVueConfig } from './uni-vue-config';